slight refactor. only show confetti when certain score threshold is met
