+++ main.js
@@ -5,7 +5,6 @@ import {
   clearHighlights,
 } from "./utils/highlight.js";
 import { createScoreGradient, displayScoreMessage } from "./utils/score.js";
-import { showTooltip, hideTooltip } from "./utils/tooltip.js";
 import { SCANNER, WEBSITE } from "./domain.js";
 
 let selection = "";
@@ -65,47 +64,6 @@ function updateButtonState(selectedButtonId) {
   // Add 'selected' class to the clicked button
   document.getElementById(selectedButtonId).classList.add("selected");
 }
-
-// Tooltip content for each info button
-const tooltips = {
-  "info-contrasting-colors":
-    "This button checks if the text on the page has sufficient color contrast for readability. It checks for at least a 4.5:1 ratio of rbg values for the background color and text color.",
-  "info-large-text":
-    "This button checks if the text size is large enough (at least size 16) for visually impaired users.",
-  "info-labeled-images":
-    "This button checks if images have labels, which improves accessibility for screen reader users.",
-};
-
-// Add event listeners to info icons
-document
-  .getElementById("info-contrasting-colors")
-  .addEventListener("mouseenter", (event) => {
-    showTooltip(event, tooltips["info-contrasting-colors"]);
-  });
-
-document
-  .getElementById("info-contrasting-colors")
-  .addEventListener("mouseleave", hideTooltip);
-
-document
-  .getElementById("info-large-text")
-  .addEventListener("mouseenter", (event) => {
-    showTooltip(event, tooltips["info-large-text"]);
-  });
-
-document
-  .getElementById("info-large-text")
-  .addEventListener("mouseleave", hideTooltip);
-
-document
-  .getElementById("info-labeled-images")
-  .addEventListener("mouseenter", (event) => {
-    showTooltip(event, tooltips["info-labeled-images"]);
-  });
-
-document
-  .getElementById("info-labeled-images")
-  .addEventListener("mouseleave", hideTooltip);
 
 //clear button
 document.getElementById("clear-button").addEventListener("click", function () {
@@ -186,7 +144,7 @@ function performScan(scanType) {
                 "visible";
               document.getElementById("score").innerHTML = data.score;
               createScoreGradient(data.score);
-              displayScoreMessage(selection, data.score);
+              displayScoreMessage(selection, data);
 
               // Show score bar
               document.getElementById("score-bar").style.visibility = "visible";
@@ -204,7 +162,6 @@ function performScan(scanType) {
                 scanType == "Contrasting Colors" ||
                 scanType == "Large Text"
               ) {
-                console.log("test console log");
                 chrome.scripting.executeScript({
                   target: { tabId: activeTab.id },
                   function: highlightInaccessibleElements,
